+++ CHANGELOG.md
@@ -9,5 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Test for pdfbox special characters : U+2002 ('enspace')
+- Module pdfbox 
 - AgeCheck with tests